add validation, update and remove user

+++ src/user/user.service.ts
@@ -1,4 +1,10 @@
-import { HttpStatus, Injectable, NotFoundException } from '@nestjs/common'
+import {
+  BadRequestException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Repository } from 'typeorm'
 import { User } from './entities/user.entity'
@@ -67,18 +73,31 @@ export class UserService {
   async updateUser(
     id: number,
     updatedUser: UpdateUserDto,
+    reqEmail: string
   ): Promise<UserResponse> {
     const user = await this.userRepository.findOne({ where: { id } })
 
     if (!user) {
-      throw new NotFoundException('User do not exist!')
+      throw new NotFoundException('User does not exist!')
+    }
+    if (user.email !== reqEmail) {
+      throw new UnauthorizedException('You can only update your own profile')
+    }
+    if (updatedUser.email) {
+      throw new BadRequestException('User cannot update their email.')
+    }
+    if (updatedUser.name) {
+      user.name = updatedUser.name
+    }
+    if (updatedUser.password) {
+      const hashPassword = bcrypt.hashSync(
+        updatedUser.password,
+        bcrypt.genSaltSync(10),
+      )
+      user.password = hashPassword
     }
 
-    await this.userRepository.update(id, updatedUser)
-
-    const newlyUpdatedUser = await this.userRepository.findOne({
-      where: { id },
-    })
+    const newlyUpdatedUser = await this.userRepository.save(user)
 
     return {
       status_code: HttpStatus.OK,
@@ -89,12 +108,16 @@ export class UserService {
     }
   }
 
-  async removeUser(id: number): Promise<DeletedUserResponse> {
+  async removeUser(id: number, email: string): Promise<DeletedUserResponse> {
     const user = await this.userRepository.findOne({ where: { id } })
 
     if (!user) {
-      throw new NotFoundException('User do not exist!')
+      throw new NotFoundException('User does not exist')
     }
+    if (user.email !== email) {
+      throw new UnauthorizedException('You can only delete your own profile')
+    }
+
     await this.userRepository.delete(id)
 
     return {
